Move tag creation to creator.go for TimescaleDB
The tags creation code was stored in main.go while all the other
database setup code was located in creator.go. This change moves
the function there and also takes the idempotent step of DROPing
the tags table if it exists before trying to create it. This helps
in scenarios where the database is already created and may have
tags left over.