Modify data generation to have tags of any type
Changes are done to the Point interface where tags are
no longer just strings, and serializers for the different databases
are modified to serialize the tags depending on their type.
For TimescaleDB and Postgres it doesn't matter which type is it,
for InfluxDB the tag is serialized as a field.